server/requirements.txt
fastapi
uvicorn
python-dotenv
openai
aiohttp
pytest
pytest-asyncio
httpx
